requirements.txt
playwright>=1.41.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
python-dotenv>=1.0.1
openai>=1.12.0
# MCP support (Playwright MCP server integration)
mcp>=1.0.0
langchain-mcp-adapters>=0.1.0
langgraph>=0.2.0
